feat(ui): storage finder detail dialogs
